Reuse cluster in AbstractDeploymentTest
