+++ hazelcast-jet-core/src/test/java/com/hazelcast/jet/impl/deployment/ClientDeploymentTest.java
@@ -18,19 +18,12 @@ package com.hazelcast.jet.impl.deployment;
 
 import com.hazelcast.internal.util.FilteringClassLoader;
 import com.hazelcast.jet.JetInstance;
-import com.hazelcast.jet.JetTestInstanceFactory;
-import com.hazelcast.jet.config.JetClientConfig;
 import com.hazelcast.jet.config.JetConfig;
-import com.hazelcast.jet.core.DAG;
 import com.hazelcast.test.HazelcastSerialClassRunner;
-import org.junit.After;
+import org.junit.BeforeClass;
 import org.junit.Rule;
-import org.junit.Test;
 import org.junit.rules.Timeout;
 import org.junit.runner.RunWith;
-
-import java.net.URL;
-import java.net.URLClassLoader;
 
 import static java.util.Collections.singletonList;
 
@@ -40,57 +33,17 @@ public class ClientDeploymentTest extends AbstractDeploymentTest {
     @Rule
     public final Timeout timeoutRule = Timeout.seconds(360);
 
-    private JetInstance client;
-    private JetTestInstanceFactory factory;
+    @BeforeClass
+    public static void beforeClass() {
+        JetConfig jetConfig = new JetConfig();
+        FilteringClassLoader filteringClassLoader = new FilteringClassLoader(singletonList("deployment"), null);
+        jetConfig.getHazelcastConfig().setClassLoader(filteringClassLoader);
 
-    @After
-    public void tearDown() {
-        factory.shutdownAll();
+        initializeWithClient(2, null, null);
     }
 
     @Override
     protected JetInstance getJetInstance() {
-        return client;
-    }
-
-    @Override
-    protected void createCluster() {
-        createCluster(new JetConfig(), new JetClientConfig());
-    }
-
-    private void createCluster(JetConfig jetConfig, JetClientConfig clientConfig) {
-        factory = new JetTestInstanceFactory();
-
-        if (jetConfig == null) {
-            jetConfig = new JetConfig();
-        }
-        FilteringClassLoader filteringClassLoader = new FilteringClassLoader(singletonList("deployment"), null);
-        jetConfig.getHazelcastConfig().setClassLoader(filteringClassLoader);
-        factory.newMember(jetConfig);
-
-        client = factory.newClient(clientConfig);
-    }
-
-    @Test
-    public void when_classAddedUsingUcd_then_visibleToJet() throws Exception {
-        URL classUrl = this.getClass().getResource("/cp1/");
-        URLClassLoader urlClassLoader = new URLClassLoader(new URL[]{classUrl}, null);
-        Class<?> personClz = urlClassLoader.loadClass("com.sample.pojo.person.Person$Appereance");
-
-        JetClientConfig jetClientConfig = new JetClientConfig();
-        jetClientConfig.setClassLoader(urlClassLoader);
-        jetClientConfig.getUserCodeDeploymentConfig()
-                       .setEnabled(true)
-                       .addClass(personClz);
-
-        JetConfig jetConfig = new JetConfig();
-        jetConfig.getHazelcastConfig().getUserCodeDeploymentConfig()
-                 .setEnabled(true);
-        createCluster(jetConfig, jetClientConfig);
-
-        DAG dag = new DAG();
-        dag.newVertex("v", () -> new LoadPersonIsolated(true));
-
-        getJetInstance().newJob(dag).join();
+        return client();
     }
 }
